Track entropy at training time
It may be useful to see how entropy changes over time, since once entropy collapses the model is unlikely to learn much more.

Not sure whether this is the best way to do it; notably, when doing it this way we aren't able to track entropy on the val set, only at training.