Bump uv to >=0.6.15
We've seen substantial churn in uv.lock files because different committers use different versions of uv. This PR requires uv 0.6.15 or higher, which is not the latest version but uses the same lockfile format as the latest version uv 0.7.17. In practice there's no reason not to just use the latest version (0.7.17 as of time of writing).